Homeworks/Homeworks-20231006T095831Z-001/Homeworks/fastapi_mysql_odev/database.py
import os
from dotenv import load_dotenv
from sqlmodel import Session
from sqlmodel import create_engine

load_dotenv()  # take environment variables from .env.
SQLALCHEMY_DATABASE_URL = os.getenv('SQLALCHEMY_DATABASE_URL')
# print(SQLALCHEMY_DATABASE_URL)

engine = create_engine(SQLALCHEMY_DATABASE_URL, echo=True)


def get_session():
    with Session(engine) as session:
        yield session
