Homeworks/Homeworks-20231006T095831Z-001/Homeworks/fastapi_mysql_odev/requirements.txt
fastapi[all]==0.86.0
uvicorn[standard]==0.18.3
SQLAlchemy==1.4.35
pymysql==1.0.2
python-dotenv==0.21.0
sqlmodel==0.0.8
passlib
bcrypt